Move Video to playlist.py
